Check viewport size every frame and resize projectM if changed.
SDL's resize events are sometimes unreliable, e.g, if using the green "fullscreen" button on macOS or toggling to the borderless, non-exclusive fullscreen mode on Linux. Getting the GL drawable size is fast and also reliable, so we can simply check it on each frame. Will also take care of DPI changes.